apps/api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.30
alembic==1.13.0
psycopg2-binary==2.9.9
asyncpg==0.29.0
pydantic[email]==2.7.0
pydantic-settings==2.3.0
redis==5.0.0
celery==5.4.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
httpx==0.27.0
websockets==12.0
apscheduler==3.10.0
